requirements.txt
python-telegram-bot
SQLAlchemy
Pillow
imgkit
pyowm